New constructor and load() method
The two biggest changes here are a new constructor that takes a String instead of a File (To reduce boilerplate and allow the calling code to not depend on java.io.File), and that I've separated out the code formerly in play() that might throw an error into a separate load() function.
load() is public so that the calling code can call load() before play() and handle any exceptions load() throws. 
I also removed the isPlaying boolean and made it so isPlaying() instead uses clip.isPlaying() as a heuristic.

This class's interface has not been changed in a breaking way however, and any code that used it before should work just as well after this commit.